make pickling possible
This implements a bare-bones __reduce__/rebuild pattern
(along the lines of https://stackoverflow.com/questions/12646436/pickle-cython-class).
This is not checked for possible memory leaks, etc., but at least
doesn't break when trying to pickle.  Possible downside is the way
.compute() is called again on rebuild.  Ideally there would be a way to just
save and restore the computed state more directly, but I don't know how to
do that.